#343 Create Spin animations *created the SpinRightAnim and SpinLeftAnim animations, both of which rotate the bone and boomerang projectiles to the right and left respectively *added animator components to both projectile prefabs *added code to the ProjectileAttack method inside the Weapon script to play this animation on the projectile if the weapon name is bone or boomerang

+++ MonsterIsland/Assets/Scripts/Helper.cs
@@ -210,7 +210,7 @@ public class Helper : MonoBehaviour {
     //used to get strings Front or Back from 1 or -1 based on the side passed
     //sides can be passed in the form of armTypes (RightArm/LeftArm) or basic sides (Right/Left)
     //if side is null, returns Left or Right based on the value of scaleX
-    public static string GetAnimDirection(int scaleX, string side = null)
+    public static string GetAnimDirection(float scaleX, string side = null)
     {
         if (side == null)
         {
